moved nine segment to ui crate
