+++ deps/ui/src/views/basic/mod.rs
@@ -2,6 +2,7 @@ mod button;
 mod circle_view;
 mod image_view;
 mod label;
+mod nine_segment_image_view;
 mod scroll_view;
 mod slider;
 mod switch;
@@ -11,6 +12,7 @@ pub use button::Button;
 pub use circle_view::CircleView;
 pub use image_view::ImageView;
 pub use label::*;
+pub use nine_segment_image_view::*;
 pub use scroll_view::*;
 pub use slider::*;
 pub use switch::*;
